Created a single entry point for training transitions and emissions.
trainModels.py

Changed to a json config file
Wrote tests for canonical DNA case

TODOs:
Still need to write test for non-canonical DNA/RNA
Need to provide default values in documentation
