+++ impl/nanopore_hdp.c
@@ -233,7 +233,7 @@ void update_nhdp_from_alignment_with_filter(NanoporeHDP* nhdp, const char* align
         line_length = stList_length(tokens);
         
         if (!warned) {
-            if ((line_length != NUM_ALIGNMENT_COLS) || (line_length != NUM_ASSIGNMENT_COLS)) {
+            if ((line_length != NUM_ALIGNMENT_COLS) && (line_length != NUM_ASSIGNMENT_COLS)) {
                 fprintf(stderr, "Input format has changed from design period, HDP may receive incorrect data.\n");
                 warned = true;
                 continue;
@@ -1150,6 +1150,19 @@ static NanoporeHDP *loadNanoporeHdpFromScratch(NanoporeHdpType nHdpType, const c
                                                double leafGammaAlpha, double leafGammaBeta,
                                                double samplingGridStart, double samplingGridEnd,
                                                int64_t samplingGridLength) {
+
+    if (nHdpType == singleLevelFixedCanonical) {
+        if ((baseGamma == NULL_HYPERPARAMETER) || (leafGamma == NULL_HYPERPARAMETER)) {
+            st_errAbort("loadNanoporeHdpFromScratch: You need to provide a base gamma and leaf gamma "
+                                "for this NanoporeHdpType\n");
+        }
+
+        NanoporeHDP *nHdp = flat_hdp_model(CANONICAL_ALPHA, CANONICAL_NUBMER, kmerLength,
+                                           baseGamma, leafGamma,
+                                           samplingGridStart, samplingGridEnd, samplingGridLength, modelFile);
+
+        return nHdp;
+    }
     if (nHdpType == singleLevelFixed) {
         if ((baseGamma == NULL_HYPERPARAMETER) || (leafGamma == NULL_HYPERPARAMETER)) {
             st_errAbort("loadNanoporeHdpFromScratch: You need to provide a base gamma and leaf gamma "
