--- src/signalalign/scripts/errorCorrectionLib.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-"""Library for error correction
-"""
-from serviceCourse.sequenceTools import reverse_complement
-from serviceCourse.parsers import read_fasta
-
-
