--- src/signalalign/train/trainModels.yaml
@@ -1,72 +1,0 @@
-# SignalAlign model training config file
-samples: [
-{
-    fast5_dirs:  [/Users/andrewbailey/data/r9/modified/],
-    fofns: [],
-    positions_file:,
-    motifs:,
-    bwa_reference: /Users/andrewbailey/data/references/ecoli/ecoli_k12_mg1655.fa,
-    fw_reference:,
-    bw_reference:,
-    alignment_file: None,
-    name: canonical,
-    number_of_kmer_assignments: 10,
-    probability_threshold: 0.8,
-    kmers_from_reference: false
-},
-{
-    fast5_dirs:  [/Users/andrewbailey/data/r9/modified/],
-    fofns: [],
-    alignment_file: None,
-    bwa_reference: /Users/andrewbailey/data/references/ecoli/ecoli_k12_mg1655.fa,
-    positions_file:,
-#    motifs: [
-#        ["CCAGG", "CEAGG"],
-#        ["CCTGG", "CETGG"]],
-    motifs:,
-    fw_reference:,
-    bw_reference:,
-    name: modified,
-    number_of_kmer_assignments: 10,
-    probability_threshold: 0.8,
-    kmers_from_reference: false,
-}
-]
-output_dir: /Users/andrewbailey/data/trainSignalAlign/
-stateMachineType: threeState
-in_T_Hmm: /Users/andrewbailey/CLionProjects/nanopore-RNN/submodules/signalAlign/models/testModelR9_acegt_template.model
-in_C_Hmm: /Users/andrewbailey/CLionProjects/nanopore-RNN/submodules/signalAlign/models/testModelR9_acegt_complement.model
-templateHdp:
-complementHdp:
-train_trainsitions: true
-train_HDP_expectations: true
-train_HMM_expectations: true
-train_transitions_options: {
-    iterations: 1,
-    training_bases: 5000,
-    diagonal_expansion: ,
-    constraint_trim: ,
-}
-train_HMM_expectations_options: [{
-
-}
-]
-train_HDP_expectations_options: [{
-    hdp_type: ecoli,
-    template:,
-    complement:
-}
-]
-job_count: 2
-twoD: false
-#degenerate: cytosine
-#bulk: false
-batch: 10
-assignments: 30
-n_aligns: 1000
-n_test_alns: 1000
-assignment_threshold: 0.8
-split: false
-debug: true
-test: false
-path_to_bin: /Users/andrewbailey/CLionProjects/nanopore-RNN/submodules/signalAlign/bin
