--- tests/trainModels-config.yaml
@@ -1,32 +1,0 @@
-# SignalAlign model training config file
-output_dir: signalAlign_unittest/
-samples: [{
-    fast5_dirs: [../tests/minion_test_reads/canonical_ecoli_R9/],
-    fofns: [],
-    bwa_reference: ../tests/test_sequences/E.coli_K12.fasta,
-    positions_file:,
-    motifs:,
-    fw_reference:,
-    bw_reference:,
-    name: canonical,
-    number_of_kmer_assignments: 10,
-    probability_threshold: 0.8,
-    kmers_from_reference: false,
-    alignment_file:
-}]
-stateMachineType: threeState
-in_T_Hmm: ../models/testModelR9_5mer_acgt_template.model
-in_C_Hmm: ../models/testModelR9_5mer_acgt_complement.model
-templateHdp:
-complementHdp:
-iterations: 3
-training_bases: 10000
-job_count: 2
-diagonal_expansion:
-constraint_trim:
-twoD: true
-alignment_file:
-DEBUG:
-test:
-path_to_bin: ./
-
